// src/interfaces/StartdocDocumentInterface.ts
interface StartdocDocumentInterface {
    url_admin_startdoc: string | null;
    owner: string | null;
    subject: string | null;
    siret: string | null;
    website_url: string | null;
    paid: string | null;
    keywords: string | null;
    content_type: string | null;
    last_update: string | null;
    comment: string | null;
    docx: string | null | undefined;
    company: string | null | undefined;
}

export default StartdocDocumentInterface;